pyproject.toml
[tool.poetry]
name = "mapmaker"
version = "1.6.5"
description = ""
authors = ["David Brooks <d.brooks@auckland.ac.nz>"]
include = ["mapmaker/output/data_mapping.json"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-pptx = "^0.6.21"
numpy = "^1.24"
beziers = "^0.5.0"
requests = "^2.31.0"
pyparsing = "^3.0.9"
pyyaml = "^6.0"
rdflib = "^6.2.0"
shapely = "^2.0.0"
pyproj = "^3.4.0"
mercantile = "^1.2.1"
mbutil = "^0.3.0"
networkx = "^2.8.8"
openpyxl = "3.1.0"
tqdm = "^4.64.0"
lxml = "^4.9.1"
opencv-python-headless = "^4.6.0.66"
svglib = "^1.3.0"
skia-python = "^87.5"
transforms3d = "^0.4.1"
tinycss2 = "^1.1.1"
cssselect2 = "^0.6.0"
webcolors = "^1.12"
xmltodict = "^0.12.0"
Pyomo = "^6.2"
svgwrite = "^1.4.3"
XlsxWriter = "^3.0.3"
multiprocess = "^0.70.13"
gitpython = "^3.1.29"
giturlparse = "^0.10.0"
types-requests = "^2.28.11.2"
colormath = "^3.0.0"
pandas = "^1.5.3"
svgelements = "^1.9.1"
flatmapknowledge = {url = "https://github.com/napakalas/flatmap-knowledge/releases/download/v0.16.3/flatmapknowledge-1.6.3-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
mypy = "^0.982"
sphinx = "^6.0"
sphinx-rtd-theme = "^1.1.1"
sphinx-argparse = "^0.3.1"

[tool.poetry.scripts]
mapmaker = 'mapmaker.__main__:main'

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = [
    'beziers.cubicbezier',
    'beziers.line',
    'beziers.path',
    'beziers.point',
    'beziers.quadraticbezier',
    'beziers.segment',
    'cssselect2',
    'cv2',
    'fitz',
    'flatmapknowledge',
    'lxml',
    'mbutil',
    'mercantile',
    'multiprocess',
    'openpyxl',
    'pptx.oxml',
    'pptx',
    'pptx.dml.fill',
    'pptx.enum.shapes',
    'pptx.enum.dml',
    'pptx.enum.text',
    'pptx.oxml.ns',
    'pptx.dml.color',
    'pptx.dml.line',
    'pptx.oxml.dml.color',
    'pptx.oxml.theme',
    'pptx.oxml.shapes.autoshape',
    'pptx.oxml.shapes.groupshape',
    'pptx.oxml.shapes.shared',
    'pptx.oxml.simpletypes',
    'pptx.oxml.slide',
    'pptx.oxml.xmlchemy',
    'pptx.shapes',
    'pptx.shapes.base',
    'pptx.shapes.connector',
    'pptx.shapes.group',
    'pptx.shapes.shapetree',
    'pptx.slide',
    'pptx.util',
    'pyomo.environ',
    'pyparsing',
    'pyproj',
    'networkx',
    'skia',
    'shapely',
    'shapely.affinity',
    'shapely.errors',
    'shapely.geometry',
    'shapely.geometry.base',
    'shapely.ops',
    'shapely.prepared',
    'shapely.strtree',
    'shapely.wkt',
    'svglib.svglib',
    'svgwrite',
    'tinycss2',
    'tqdm',
    'transforms3d',
    'webcolors',
    'xlsxwriter',
    'xmltodict',
]
ignore_missing_imports = true
